+++ src/AttributeInjectors/HostnameAttributeInjector.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace Superbalist\EventPubSub\AttributeInjectors;
+
+use Superbalist\EventPubSub\AttributeInjectorInterface;
+
+class HostnameAttributeInjector implements AttributeInjectorInterface
+{
+    /**
+     * @var string
+     */
+    protected $key;
+
+    /**
+     * @param string $key
+     */
+    public function __construct($key = 'hostname')
+    {
+        $this->key = $key;
+    }
+
+    /**
+     * @return string
+     */
+    public function getAttributeKey()
+    {
+        return $this->key;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getAttributeValue()
+    {
+        return gethostname();
+    }
+}