+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0 - ?
+
+* Add `setAttribute` method to `EventInterface`
+* Add support for "attribute injectors" with `AttributeInjectorInterface`
+* Bundle `DateAttributeInjector`, `GenericAttributeInjector`, `HostnameAttributeInjector`, `Uuid4AttributeInjector`
+* Add new `addAttributeInjector` and `getAttributeInjectors` methods to `EventManager`
+
 ## 1.0.0 - 2017-01-27
 
 * Initial release